OPEN_AI_AGENTS_SDK_NEW/22_memory_management/pyproject.toml
[project]
name = "22-memory-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "mem0ai>=0.1.116",
    "openai-agents>=0.3.0",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
